# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JEE Study Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface SubjectTarget {
  id: number;
  subject: string;
  weeklyTarget: number;
  createdAt: string;
  updatedAt: string;
}

export interface SubjectTargetInput {
  subject: string;
  weeklyTarget: number;
}

export interface SubjectWeeklyProgress {
  subject: string;
  target: number;
  completed: number;
  weekStart: string;
  weekEnd: string;
}

export interface WeeklyProgressSummary {
  subjects: SubjectWeeklyProgress[];
  weekStart: string;
  weekEnd: string;
}

export interface ErrorStat {
  /** YYYY-MM */
  month: string;
  conceptualError: number;
  calculationMistake: number;
  unknown: number;
  readingError: number;
  total: number;
}

export interface DailyQuote {
  text: string;
  author: string;
  date: string;
}

export interface TestAnalysis {
  id: number;
  title: string;
  /** physics | chemistry | maths | mixed */
  subject: string;
  date: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  maxScore?: number | null;
  /**
     * Minutes
     * @nullable
     */
  timeTaken?: number | null;
  revised: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TestInput {
  title: string;
  subject: string;
  date: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  maxScore?: number | null;
  /** @nullable */
  timeTaken?: number | null;
  revised?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface TestUpdate {
  title?: string;
  subject?: string;
  date?: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  maxScore?: number | null;
  /** @nullable */
  timeTaken?: number | null;
  revised?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface WrongQuestion {
  id: number;
  testId: number;
  questionText: string;
  conceptualError: boolean;
  calculationMistake: boolean;
  unknown: boolean;
  readingError: boolean;
  takeaway: string;
  createdAt: string;
}

export interface WrongQuestionInput {
  questionText?: string;
  conceptualError?: boolean;
  calculationMistake?: boolean;
  unknown?: boolean;
  readingError?: boolean;
  takeaway?: string;
}

export interface WrongQuestionUpdate {
  questionText?: string;
  conceptualError?: boolean;
  calculationMistake?: boolean;
  unknown?: boolean;
  readingError?: boolean;
  takeaway?: string;
}

export interface HealthStatus {
  status: string;
}

export interface ProgressItem {
  id: number;
  /** physics | chemistry | maths */
  subject: string;
  /** dpp | jee_replica | practice_sheet | chapterwise_test | cengage | exercise_book */
  category: string;
  /**
     * For chapterwise_test: part1..part5
     * @nullable
     */
  subCategory?: string | null;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  totalQuestions?: number | null;
  /** @nullable */
  completedQuestions?: number | null;
  /**
     * Score as percentage 0-100
     * @nullable
     */
  score?: number | null;
  /** not_started | in_progress | completed */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProgressInput {
  subject: string;
  category: string;
  /** @nullable */
  subCategory?: string | null;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  totalQuestions?: number | null;
  /** @nullable */
  completedQuestions?: number | null;
  /** @nullable */
  score?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface ProgressUpdate {
  /** @nullable */
  subCategory?: string | null;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  totalQuestions?: number | null;
  /** @nullable */
  completedQuestions?: number | null;
  /** @nullable */
  score?: number | null;
  status?: string;
  /** @nullable */
  notes?: string | null;
}

export interface CategorySummary {
  category: string;
  totalItems: number;
  completedItems: number;
  /** @nullable */
  avgScore?: number | null;
}

export interface SubjectSummary {
  subject: string;
  totalItems: number;
  completedItems: number;
  inProgressItems: number;
  /** @nullable */
  avgScore: number | null;
  byCategory?: CategorySummary[];
}

export interface ProgressSummary {
  subjects: SubjectSummary[];
  overallCompleted: number;
  overallTotal: number;
}

export interface CustomTask {
  id: string;
  label: string;
  done: boolean;
}

export interface DailyLog {
  id: number;
  /** ISO date string YYYY-MM-DD */
  date: string;
  lectures: boolean;
  dpp: boolean;
  revision: boolean;
  homework: boolean;
  customTasks?: CustomTask[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface DailyLogInput {
  date: string;
  lectures?: boolean;
  dpp?: boolean;
  revision?: boolean;
  homework?: boolean;
  customTasks?: CustomTask[];
  /** @nullable */
  notes?: string | null;
}

export interface DailyLogUpdate {
  lectures?: boolean;
  dpp?: boolean;
  revision?: boolean;
  homework?: boolean;
  customTasks?: CustomTask[];
  /** @nullable */
  notes?: string | null;
}

export interface DayCompletion {
  date: string;
  completedCount: number;
  totalCount: number;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  totalDaysLogged: number;
  last30DaysCompletion: DayCompletion[];
}

export type ListProgressParams = {
subject?: string;
category?: string;
};

export type ListDailyLogsParams = {
from?: string;
to?: string;
};